Force units of SpeciesConc collection to v/v dry for GEOS-Chem classic simulations
To avoid issues with the GEOS-Chem restart file, which always
expects units of v/v dry for species concentrations, the units of
the SpeciesConc collection units are forced to v/v dry. Changing
units, by changing the metadata in state_diag_mod.F90 is still
allowed when running in GCHP or in GEOS-5 (i.e. when compiling with
ESMF=y).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>