Add error handling for if HEMCO max categories is exceeded
Each emission can have multiple categories listed in HEMCO_Config.rc
up to a maximum set as a parameter in HEMCO/Core/hco_datacont_mod.F90.
Previously if this maximum was exceeded then HEMCO would crash later on
in the run without a useful error. Now HEMCO exits gracefully with an
explanatory message if the number of categories parsed in
HEMCO_Config.rc exceeds the maximum.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>