Add check to input.geos to make sure simulation is longer than defined timesteps
To avoid issues when saving out netCDF diagnostics via the history
component, we now check to make sure that the simulation length is not
shorter than the transport and/or chemistry timesteps. This should only
impact short, test simulations which should be run for at least an entire
transport and chemistry timestep anyway.