Rename restart field in State_Diag to SpeciesRst
Save out species concentrations to the SpeciesRst_* fields specified in
HISTORY.rc and stored in State_Diag%SpeciesRst. We were previously saving
those fields out to SpeciesConc_*, but this causes problems in GCHP because
internal state fields can't have the same name as fields in history.
To pass State_Chm%Species values to State_Diag%SpeciesRst, an additional
call to Set_SpcConc_Diagnostic needed to be added to diagnostics_mod.F90.
In that routine, we also restored the ability to change units of the
SpeciesConc collection for all simulations and instead only force the units
to v/v dry when the field is 'SpeciesRst'.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>