Add fix for saving ND21 diagnostic to multiple levels
In an earlier commit for interfacing GEOS-Chem into GEOS, the code

  LD21 = MIN( Input_Opt%ND21, LLCHEM )

was replaced with

  Input_Opt%LD21 = MIN( Input_Opt%ND21, LLCHEM )

This meant LD21 was never changed and remained at the initialized
value of 1. When writing to bpch file in diag3.F, LD21 is used,
and this caused the optical depth diagnostics to only be saved out
for one level. To fix this, we just set LD21 = Input_Opt%LD21.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>