Add more reliable error trap for missing restart or initial met files
Added an error trap after the first call to EMISSIONS_RUN in
GeosCore/main.F (for GEOS-Chem "Classic" only).  This will print an
informational error message instructing the user to doublecheck for
missing/invalid restart files or met field files that correspond
to the start date of the simulation.

Prior to this update, a missing restart file would not be trapped, and
the run would make it into PBL mixing, where it would die with a
"bad sum at I,J,L" error.

Also now word wrap the instructions message in routine Error_Stop
(in GeosCore/error_mod.F).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>