Add flexgrid_read_mod.o to list of dependencies for hcoi_gc_main_mod.o
This is a minor fix for an error in compiling GEOS-Chem Classic that we noticed when compiling GEOS-Chem Classic with a parallel compilation option.
The compile process fails occasionally due to the Get_Met_Fields subroutine in HCOI_GC_Main_Mod requiring FlexGrid_Read_Mod in a use statement,
but this module was not included in the Makefile dependencies. This is fixed by this commit.

Signed-off-by: Jimmie Lin <jimmie.lin@gmail.com>