Add long name to passive species menu
Users now have the option of including a long name string (use underscores,
not spaces!) for entries in the passive species menu in input.geos.
Eventually, we would like to pass this long name to the History component
so that it is saved out to the long_name attribute in the netCDF files
output from GEOS-Chem. For now, this is useful for clarity in the log
file and input.geos file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>