Remove references to clock species from UCX code
In 12.2.0, a clock tracer is added to the passive species menu
for the Rn-Pb-Be simulations. This tracer is one of many that will
be used to compare GEOS-Chem to the GEOS model. To avoid confusion,
we remove reference to the clock species from ucx_mod.F. This
species is not defined or used anywhere else and was likely added
for testing purposes.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>