Add code to update TR_Clock species from common set of GEOS-GC species
A clock tracer is added to the passive species menu for Rn-Pb-Be
simulations (and may easily be added to other simulations). This
tracer is one of many that will be used to compare GEOS-Chem to
the GEOS model. The GMI tracer mechanism used by GMAO defines the
clock tracer as

   Used for looking at residual circulation and mixing.

   The model time step (in days) is added to every grid box every
   time step with a total sink at the surface every time step.

NOTE: The units of the clock species in GEOS-Chem will be in seconds,
not days.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>