Added parameterization to GeosCore
This parameterization, based on Sakamoto et al (2016), estimates the amount of
near-source, sub-grid scale coagulation happening in a biomass burning plume.
Can be turned on or off. When on, the default assumption is that each smoke plume
is completely seperate from the others (i.e. there is no overlap of the plumes).
There is also an option for all smoke plumes in a grid box to overlap completely.
When being used, this parameterization changes the median diameter and modal width
of biomass burning emissions to account for coagulation.