Allow users to specify TAU in e-fold time for passive species (default is half-life)
To use e-fold time instead of the half-life for TAU in the passive species
menu, users just need to make sure the species name begins with 'TR_e'.
For example, TR_e90, TR_e90_N, and TR_e90_S in the new tracer simulation
all specify the loss (90 days) in e-fold time. By default, TAU is assumed
to be the half-life.

This capability was adapted from the code in TR_GridCompMod.F90 from the
GEOS model.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>