Fix bug in stratosperic aerosol surface area netcdf diagnostics
This update was submitted by Jonathan Moch who pointed out that "As
written, the code is filling the stratospheric liquid aerosol surface
area density and the polar strat cloud surface area with the
stratospheric liquid aerosol surface area and then refilling the values
with polar strat surface area. So if you aren't running the model inside
GEOS-5 the AerSurfAreaPolarStratCloud and AerSurfAreaStratLiquid are
identical, which shouldn't be the case."

A fix for this was implemented by Christoph Keller in the version of
GEOS-Chem used in GEOS-5. To fix the issue in the offline model we
use the same solution.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>