Add boundary condition collection to history component
A BoundaryCondition collection has been added to HISTORY.rc for saving
species concentrations from a global simulation for use as boundary
conditions in a nested grid simulation. The code updates for that new
collection are added in this commit.

  #==============================================================================
  # %%%%% THE BoundaryCondition COLLECTION %%%%%
  #
  # GEOS-Chem restart file fields
  #
  # Available for all simulations
  #==============================================================================
    BoundaryCondition.template:  '%y4%m2%d2_%h2%n2z.nc4',
    BoundaryCondition.format:    'CFIO',
    BoundaryCondition.frequency: 00000000 030000
    BoundaryCondition.duration:  00000001 000000
    BoundaryCondition.mode:      'instantaneous'
    BoundaryCondition.fields:    'SpeciesBC_?ALL?              ', 'GIGCchem',
  ::

The capability to read these fields from netCDF files via HEMCO will be
added in a future commit.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>