Move clock tracer to a field in State_Met
The clock tracer (TR_Clock) should be in units seconds, so including it
in State_Chm%Species doesn't make sense because the units are expected to
be in kg/kg dry and are converted often. To avoid introducing numerical
noise due to unit conversions and confusion with the units of the clock
tracer, we will instead include a field "Clock" in State_Met to track the
age of air in seconds. The diagnostic can be saved out via the StateMet
collection in HISTORY.rc.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>