Wrap remaining HEMCO extensions in instances
In HEMCO 2.1.009, several extensions were wrapped in instances. The GC_RnPbBe
extension was not updated. This resulted in zero emissions for Rn and Be
when using any version of IFORT. When using gfortran, Rn and Be emissions
seemed to be fine. This is now fixed by wrapping the GC_RnPbBe extension
in instances.

The following extensions are now also wrapped in instances: GC_POPs,
DustGinoux, and TOMAS_Jeagle.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>