Add routine to save the GC species database to JSON format
In Headers/species_database_mod.F90, we have added routine
SpcData_To_JSON, which saves several fields of the GEOS-Chem Species
Database to JSON format.  This will make it easier for Python code
(such as GCPy) to obtain species properties that are necessary for
either regridding or for unit conversion.

Routine Species_To_JSON is called from Init_Species_Database.  It is
currently commented out by default.  Perhaps in the future we can add
a logical to Input_Opt which would toggle calling this routine.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>