Prevent repeat allocation of dry deposition module variables.
This is a minor code update for INIT_DRYDEP to skip variable allocation
if it has already been allocated once. This is correct because all the
grid-dependent variables in DRYDEP_MOD have been moved to State_Chm,
thus the dry deposition module only contains data generic to the species
list, which is consistent in a given run and does not need to be read
more than once.

This scenario is only possible in the multiple domain configuration of
WRF-GC model, and does not affect other GEOS-Chem configurations at this
time.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>