Add error checking to Compute_Olson_Landmap routine
We now throw an error if the maxFracInd variable is less than 1.
This indicates a potential error with the regridding (i.e. a grid
box does not have at least one land type).  This can happen if there
is a typo in the LANDTYPE* entries in the HEMCO_Config.rc file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>