Revert "Print out entire species name string to tracerinfo.dat"
This reverts commit d2a8cd1445a3456df008729ce23d066d6edc328c.

This was done to avoid issues with xbpch which apparently depends on
the tracerinfo.dat format before this update. Reverting to 8 characters
will cut off some tracer names in the bpch output for Transport
Tracers simulation, but this can be avoided using the netCDF diagnostics
instead or by hacking the GAMAP routines.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>