Add an improved test for setting interval alarms
Rewrote the algorithm to compute the Update Interval, File Write
Interval, and File Close Interval.  Intervals of longer than 1 month
or 1 year are now allowed.  Also, the interval computation should be
much more robust when straddling leap year days.

This update fixes an issue with the Restart collection (when used with
the "End" frequency) not properly writing out files for intervals that
straddle leap year days.

Also note: the computation for deltaYMD and deltaHMS in history_mod.F90
(which is used for the "End" frequency) has also been updated to be
more robust.  The new algorithm should work for most common model
simualtion intervals (i.e. integral numbers of minutes or hours or days
or months).  There might be a few edge cases that would cause it to
fail, but the new algorithm is definitely an improvement over the
prior situation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>