Add fixes in hcoi_gc_diagn_mod.F90 to get SO2 volcano emissions from AeroCom_Volcano extension
The volcano emissions previously came into HEMCO via the base emissions, but
are now processed through the AeroCom_Volcano extension. This needed to be
updated in hcoi_gc_diagn_mod.F90 in order to save out these emissions to BPCH.

In hcox_aerocom_mod.F90, we also no longer need to hardcode year to
the 2005-2012 time range because the OMI-based emissions are available
for 1979-present.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>