Revert to getting volcanic emissions from base emissions for BPCH diagnostics
In the original 12.3.0 benchmark, we updated to OMI-based volcanic
emissions that were obtained from the AeroCom_Volcano extension.
Those emissions proved to be too low, so we have reverted to the
old volcanic emissions and therefore need to get those emissions from
the base emissions (ExtNr=0) in hcoi_gc_diagn_mod.F90.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>