Add "RFY" time flag option in HEMCO
The Y option had been added to several other time flag options in
HEMCO_Config.rc in an earlier commit to ensure that the simulation
year is always used for met fields and restart file fields. Here,
we also add the Y option to the "RF" flag. "RFY" will be used by
default for met fields to force met fields to be within the specified
range but also to avoid letting the emissions year override the
simulation year for those fields.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>