Now dimension State_Met%Area_M2 as (I,J) instead of (I,J,1)
The State_Met%Area_M2 field was allocated as (I,J,1), but this causes
issues with GCHP.  When GCHP sees a 3rd dimension, it will try to
save diagnostics with all levels.

The quick fix is to allocate State_Met%Area_M2 as an (I,J) array
instead of as an (I,J,1) array.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>