Now loop over (I,J) when computing State_Met%IsLocalNoon
The State_Met%IsLocNoon(I,J) and State_Met%LocTime(I,J) fields
were depend on local solar time, which depends on longitude.

For GEOS-Chem "Classic", longitude only depends on the I index,
and we called GET_LOCALTIME(I,1,1).

But for GCHP, which uses cubed-sphere grids, longitude is a function
of (I,J), so we have to loop over (I,J) and call GET_LOCALTIME(I,J,1)
when computing the State_Met%LocalTime and State_Met%IsLocalNoon
fields.

This is necessary to make the local-noon J-value diagnostic in GCHP
more closely match that in GC "Classic".

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>