Update State_Grid%GridRes strings to always use real numbers and not integers
For the resolution strings in input.geos to be properly converted to
State_Grid%DX and State_Grid%DY, we need to pass real numbers (e.g. 4.0x5.0)
instead of integers (e.g. 4x5). This change was made in an earlier commit,
but uses of State_Grid%GridRes throughout the code were not updated.
That is fixed in this commit.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>