Use FAST-JX directory provided in input.geos
Use the directory in input.geos' photolysis menu as the directory to look for FAST-JX data files.
The default setting is to use the files in /ExtData/CHEM_INPUTS/FAST_JX/ for most recent version (e.g. subfolder v2018-09).
However, a user can now set the FAST-JX directory to be a directory of their choice (e.g. a model's run directory).

Signed-off-by: Tomas Sherwen <tomas.sherwen@york.ac.uk>