Add fix for calculation of YEdge, YEdge_R, and YSin at the North Pole
The YEdge value for the last latitude edge is hardcoded to -90 (taken
from an earlier version of gc_grid_mod.F90). However, we need to make
sure the second-to-last latitude edge values are adjusted by recomputing
them after the last latitude edges are computed. With this fix, the YSIN
and Area_M2 values are now identical between the previous stable version
of GEOS-Chem (12.3.2).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>