Aerosol surface area from ISORROPIA water content
Use the aerosol water content from ISORROPIA II to
infer the volume and surface area of sulfate-nitrate-ammonium
aerosol, which is used in the heterogeneous chemistry.

The aerosol optics used for FastJ and RRTMG continue to use lookup
tables for hygroscopic growth (differ from ISORROPIA) because
all of the optical properties (phase function, index of refraction)
must be consistent with the composition and size.

Signed-off-by: Christopher D. Holmes <cdholmes@fsu.edu>