Add fixes for FlexGrid in GCHP
In gc_grid_mod.F90 combined routines Compute_Grid and DoGridComputation
into a single routine. Compute_Grid was made to a wrapped for the
original HEMCO implementation but doesn't appear to be needed anymore.
To simplify things, we moved the grid computation code into Compute_Grid
and removed DoGridComputation completely.

This code has also been blocked off by CPP switches to ensure it is not
called when running GEOS-Chem in an ESM or in GCHP mode. The grid does
not need to be computed in those cases.

Finally, updated routines SetGridFromCtr and SetGridFromCtrEdges to use
NX and NY directly from State_Grid instead of passing those values as
input arguments.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>