Temporarily add Area_M2 back to State_Met to avoid issues in GCHP benchmarking
The Area_M2 field is now in State_Grid, however to avoid breaking GCHP
benchmarking we still copy this field to State_Met%Area_M2. This is required
to save Area out to the StateMet collection via HISTORY.rc and is used
in GCPy to compute things like emissions totals. Eventually, Area_M2
should be removed from State_Met. We may be able to obtain that information
from ESMPy instead.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>