Add bug fix for History collections that use .filename instead of .template
In HISTORY.rc the filename attribute can be used to define the complete
path and filename of the output file. When filename is specified, the
template attribute may be ommitted. To avoid errors when reading the
HISTORY.rc files, add '.filename' as a possible way of identifying a
new collection section.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>