Add fix to DEPVEL to avoid array-out-of-bounds errors in TOMAS simulations
The TOMAS simulations now include the bulk dust species (DST1-4). The TOMAS-
specific code in routine DEPVEL for determining BIN number did not account
for these species and trying to compute BIN for those species would result
in an array-out-of-bounds error. A fix has been added to compute DIAM and
DEN using A_RADI and A_DEN for the DST1-4 species and to use SIZ_DIA and
SIZ_DEN for the TOMAS aerosols.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>