Bug fix: APM variable MODISOUT does not need to be held PRIVATE
In the APM driver module (GeosCore/apm_driv_mod.F), the variable
MODISOUT was added to an !$OMP PRIVATE list.  But MODISOUT does not
need to be private since it has the same scope as the parallel loop.
This was causing unit tests to fail with floating-point errors.

The fix is to remove MODSIOUT from the !$OMP PRIVATE statement.
APM simulations now pass unit tests.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>