Modifications to enable the ConcAboveSfc diagnostic collection
Added several modifications in various modules to enable the
ConcAboveSfc diagnostic.  This is the concentration of O3 and HNO3
at a user-defined altitude above the surface.

Long story short: You have to turn on the ConcAboveSfc collection
in HISTORY.rc and also set the "Diag alt above sfc [m]" line in
input.geos to the desired altitude.

NOTE: We have also tagged O3 and HNO3 with the "DRYALT" tag in
the species database.  These are the only two species for which
this diagnostic is appropriate.

While most of the machinery has been done, the actual computation
in diagnostics_mod.F90 still needs to be added.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>