Rename DryDepZLALT1 diagnostic to DryDepZLfrac and make it a counter
We now have renamed the State_Diag%DryDepZLALT1 diagnostic to
State_Diag%DryDepZLfrac.  This diagnostic will be assigned a value of
either 0 (where the M-O similarity value z/L > 1) or 1 otherwise.

Because the conversion from the surface layer = ~60m to 10m (or whatever
altitude is selected by the user) only holds where z/L <= 1, we will
need to update SpeciesConc10m only where z/L <=1, so that it in
essence becomes a cumulative sum.  Then we will need to divide the
SpeciesConc10m by DryDepZLfrac in post-processing (similar to how
JNoon is divided by JNoonFrac in post-processing).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>