Take the min & max level of the COLLECTION.levels in HISTORY.rc
In GCHP with the new MAPL v1.0.0, you can restrict the number of levels
that get archived to the netCDF ("History") diagnostics by listing
the levels that you want to save out under the COLLECTION.levels tag.

We have now made GEOS-Chem "Classic" use the same format in the
HISTORY.rc file.  But because it is difficult to point to array
sub-slices that are not contiguous, we will archive all the min and
max values listed in COLLECTION.levels.

For example, if you type:

    SpeciesConc.levels:  2 5 10 14

then GEOS-Chem "Classic" History would save model levels 2-14 to
the SpeciesConc collection netCDF files, etc.

We have also now made sure that all module variables in history_mod.F90
get deallocated.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>