Replace species conc unit conversions for budget with in-place computation
This update removes several conversions of State_Chm%Species from kg/kg
to kg and back again that occur every timestep if budget diagnostics
are turned on. The update introduces numerical noise differences due to
the removal of multiplication and division operations on species
concentrations. It also speeds up the model when budget diagnostics are
turned on.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>