Add DO_PHOTCHEM logical back to flexchem_mod.F90
The switch to turn photolysis chemistry on/off has been added back to
flexchem_mod.F90. DO_PHOTCHEM should only be set to FALSE for testing
and debugging purposes and will require recompiling GEOS-Chem. The switch
has been reinstated so that FAST-JX is still called, photolysis rates are
still computed and passed to diagnostics, but photolysis rates are not
passed to KPP's PHOTOL array and therefore do not impact chemistry.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>