Wrap GLOB_RCONST array in flexchem_mod.F90 with #ifdef MODEL_GEOS
The GLOB_RCONST array is only used when connecting GEOS-Chem to the
NASA GEOS ESM.  But there was an instance of the GLOB_RCONST array
being updated even for GCC and GCHP simulations.

Therfore, we have now wrapped the line where GLOB_RCONST gets updated
with an #if defined( MODEL_GEOS ) block.  This will avoid using excess
CPU cycles.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>