Added fixes to get RRTMG code to build with gfortran
The gfortran compiler does not like instances of e.g.

   IF ( INTROP(I,J,L) .eq. TRUE ) THEN

because it wants you to use the .eqv. statement to compare LOGICAL
variables.  But this is unnecessary, as you can just use the

   IF ( INTROP(I,J,L) ) THEN

Because INTROP is just copied from State_Met%InTroposphere, we can
just replace this with:

   IF ( State_Met%InTroposphere(I,J,L) ) THEN

Doing this resulted in a successful compile with GNU Fortran 9.2
(Spack-built libraries).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>