Force setting of PD64=W_ to get printout of metadata in tracerinfo.dat
Due to some rearrangement of code recently, the PD64 bpch parameter
was not getting set properly.  We now force it to PD64=W_ so that
the ND64 metadata will be saved out to tracerinfo.dat

NOTE: We are going to start removing the bpch diagnostics in 12.7.0,
but this is a necessary step in order to compare ND64 bpch and
netCDF diagnostics.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>