Initial modifications for removal of most bpch diagnostics
We have started to remove most of the bpch diagnostics from the
code everywhere.  This is the first round of updates.

These updates were validated with difftests for CH4 and aerosol-only
simulations.  More to be done.

We will preserve the following diagnostics for the time being:
(1) TOMAS: ND06, ND44, ND59-61.
(2) RRTMG: ND72
(3) Timeseries: ND51 and ND51b

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>