Additional modfications to remove bpch diagnostics
We have continued to remove the bpch diagnostics from the code.

The only ones that will be preserved are:

* For MODEL_GEOS: ND21 (why is this???
* For RRTMG: ND72
* For TOMAS: ND06, ND44, ND59, ND60, ND61
* For Hg: ND03
* For POPS: ND53
* For Lightning Flashrates: ND56 -- will soon be converted to netCDF
* Satellite timeseries: ND51, ND51b

Also note: there seems to be a weird situation in vdiff_mod.F90,
where the quick fix for mass conservation by Jintai Lin was placed
inside an IF ( ND15 > 0 ) block.  Apparently this never gets
executed when BPCH_DIAG=n.  Need to investigate further.

This update was validated with a Hg difftest, with BPCH=DIAG=y.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>