Move debug printout switch (aka ND70) to the SIMULATION MENU for GCHP
This update was originally applied to GEOS-Chem Classic by Bob Yantosca:
"The switch that toggles printing of debug output (formerly ND70) was
contained in the DIAGNOSTIC MENU of the input.geos* files.  However,
the DIAGNOSTIC MENU will now only be read if GEOS-Chem is compiled
with BPCH_DIAG=y. Therefore, we need to move the debug output switch
from the DIAGNOSTIC MENU to the SIMULATION MENU, which is always read."

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>