Update missing GIGC.mk compile error message to be more clear
If GIGC.mk cannot be found it is because the GCHP repository is not
present or is misnamed. The error message now suggests that the user
check these things. This error should only be triggered if a user
is compiling GCHP and does not have the GCHP repo subdirectory.
Moving the run directory creation to the GEOS-Chem repository makes
this user error more likely to occur.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>