Use REAL8 not 4 for DELP_DRY in GCHP restart(removes multi vs single diffs)
State_Met%DELP_DRY is carried as REAL8 but was written to the GCHP
restart file via the MAPL internal state as REAL4. This resulted in
small number differences in species concentrations when breaking up
a run into small consecutive runs. Storing in the internal state as REAL8
removes the differences.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>