Fix compilation error for gfortran 9.2.1+ in hco_config_mod.F90
See Github issue:
  https://github.com/geoschem/geos-chem/issues/113

Long story short: We had a couple of instances of a loop index
variable (N) trying to be modified within a do loop in hco_config_mod.F90.
Previous versions os ifort and gfortran let this go (or maybe would
fix it under the hood).  But Gfortran 9.2.1 flagged this as an error.

The simple solution is to just rename the variables that are being
modified to something other than N.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>