Restore original filename when file not found in HEMCO's SrcFile_Parse
When a file is not found for a given date and the cycle flag is enabled,
HEMCO will step backwards in time by default to look for a file. When the
maximum number of timesteps is reached, HEMCO stops looking. This often
leads to error messages saying that a file is not found for year 1715,
which can be confusing to users. To avoid this confusion, when a file is
not found after looping in time, the original filename will be restored
for printing to HEMCO.log.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>