Added a missing OPEN statement in UCX routine SFCMR_READ
For some reason, there is no OPEN statement in the UCX routine
SFCMR_READ, which is where GCHP and GEOS-5 read in the various
UCX surface mixing ratios from ASCII files.  Not having an OPEN
statement was causing the code to try to read from a file called
fort.14 (i.e. where 14 is the unit number).

We have now added an OPEN statement.  We have also updated some of
the READ commands to use the '(a)' format when reading in a blank
line.  This should be safer for the newer compilers.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>