Remove GCHP output file gcchem_internal_checkpoint if present at run start
Update all sample run scripts to ensure a restart file with the original
name used by MAPL is not present in the run directory prior to executing
geos. Its presence will cause GCHP runs to fail. The output restart file
is renamed after a successful run within the run scripts, but a file with
the original name may persist if the previous run was unsuccessful,
causing the run to crash.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>