Make sure the HEMCO standalone gets the met field and grid resolution
In one of the recent updates for FlexGrid, the settings
HcoState%Config%MetField and HcoState%Config%GridRes are manually
set from input.geos when running HEMCO connected to GEOS-Chem, GCHP,
or another parent model.  But these fields were undefined when
running HEMCO in standalone mode.  This error was probably introduced
during the updates for FlexGrid.

We have added some extra code in Core/hco_config_mod.F90 so that
we look for the MET and RES fields from the HEMCO_sa_Config.rc
(or HEMCO_Config.rc) file, and then use these to set the values of
HcoState%Config%MetField and HcoState%Config%GridRes.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>