Automatically include subrtn Get_OHreactivity in Standard mechanism build
Previously every time the Standard mechanism was built with KPP the
python script OHreact_parser.py had to be run manually. That script output
a text file containing subroutine Get_OHreactivity that then had to be
manually copied to gckpp_Util.F90. With this update the script to
build the mechanism (build_mechanism.sh) calls the python script and
the python script is updated to write directly to gckpp_Util.F90.
Manual updates are no longer needs.

The chemistry in the Standard mechanism does not change with this update
but the time and working directory included in the files is updated for
a more recent build. The OHreact equation in Get_OHreactivity does
change with this commit because the version of the subroutine previously
included was outdated. It is now up-to-date with the latest mechanism
updates.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>