Insert rxn rate diagnostic code in KPP when building Standard mechanism
Reaction rate diagnostics are currently implemented for GEOS-Chem if
running in GEOS. Code needed in gckpp_Function.F90 had to be manually
added after building the mechanism for the diagnostics to be functional.
With this update the code is inserted automatically from within the
build_mechanism.sh script. The code itself is also added back into
gckpp_Function.F90 (it was removed in the last commit). The diagnostic
is still GEOS-only within this update. No other parts of the mechanism
are changed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>