Update some HEMCO extensions for the dry-run simulation
We now print files read by the Volcano, Paranox, and Lightning NOx
extensions to the GEOS-Chem dry-run log file.

We also now have simplified the arguments.  Any of:

 ./geos --dry-run
 ./geos --dryrun
 ./geos -D

will turn on the dryrun and send output to GEOSChem.DryRun.log.
To change the logfile name, add a second argument, such as:

 ./geos --dry-run dryrun.log

etc.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>