Modifications to make the dry-run output more consistent
Updated the messages for FAST-JX and HEMCO extensions so that the
outputs to the log file are more consistent.  We also now add error
checks so that the non-dry-run simulations will exit immediately if
each requested file is not found.

Also note: We need to enter the HEMCO extension run routine so that
we can print out the names of the volcano files, but prevent other
routines from executing.  This will be done in a subsequent commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>