Add OH reactivity diagnostic
The OH reactivity is computed by parsing the KPP mechanism and combining
all reactions that consume OH into a single ‘rate’, which is the OH
consumption. Christoph Keller originally created the diagnostic for use
in GEOS. With this update it is now available to all users of GEOS-Chem.

This diagnostic is not included in HISTORY.rc by default. To enable this
diagnostic add OHreactivity to any diagnostic collection.