Add simple version of KPP equation reaction rate diagnostics
This first version of the KPP equation reaction rate diagnostics for
non-GEOS GEOS-Chem runs allocates reaction rate diagnostics for ALL
equations as long as a diagnostic for one equation is specified in
HISTORY.rc. This will be updated to allocated based on content of
HISTORY.rc in a future update to reduce memory associated with these
diagnostics.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>