Bug fix: allow OHreact_parser.py to work on any number of reactions
Previously '&' and a newline were written after every five reactions
to gckpp_Util.sh when using OHreact_parser.py to construct subroutine
get_OHreactivity. If the total number of reaction terms written was
divisible to 5, however, as is the case with the SOA_SVPOA mechanism,
then compilation would fail due to the '&' at the end of the last
line in the equation. This fix moves the write of '&' and a newline
to higher up in the loop, prior to finding a new qualifying OH
reaction, so that it only happens when at least one more equation
term will be written.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>