Remove dry-run log file; all dry-run output now goes to stdout
We have removed the DryRunLog and DryRunLUN variables from Input_Opt
and HcoState%Options.  The GEOS-Chem and HEMCO standalone dry-run
simulations now write all file status output to stdout (which can
be redirected to a log file).

The log file output must now be parsed by a script to produce a unique
list of files that are required, or a set of download commands.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>