Add tagged diagnostic list module for storing HISTORY.rc tag summaries
Taggeddiaglist_mod.F90 is similar in form and function to diaglist_mod.F90.
A TaggedDiagList object is created in main.F to store tag summary per
tagged state_diag array that is enabled in config file HISTORY.rc. If
debug prints are on then the summary of the contents within the
TaggedDiagList object are printed to stdout, with the following format:

 =============================
 Summary of tagged diagnostics

 SPECIESRST
    isWildcard:  T
  numWildcards:  1
                   ALL
       numTags:  0

 DRYDEP
    isWildcard:  T
  numWildcards:  1
                   DRY
       numTags:  4
                   O3
                   DST1
                   DST2
                   SO2

 RXNRATE
    isWildcard:  F
  numWildcards:  0
       numTags:  2
                   EQ001
                   EQ200

The information in taggeddiaglist is not used in the current commit,
but it could be used in the future to initialize state_chm arrays
to only tags requested for output, and to set the index mapping for
the reduced dimensionality.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>