Move the call to Init_Dry_Run before Read_Input_File in main.F
We have to test if the run is a dry-run simulation before reading
the input.geos file.  Otherwise the run will try to validate the
existence of several directories and will choke if they are not
found.

In routine VALIDATE_DIRECTORIES in input_mod.F, we exit if it is
a dry-run simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>