Add updates to read met fields daily instead of hourly
The HEMCO configuration file has been updated to use hour input * instead of
0-23, telling HEMCO to read all hours at the start of a new day. The code
in flexgrid_read_mod.F90 has been updated to determine the proper time index
for each met field and pass that to HEMCO via the Get_Met_2D and Get_Med_3D
routines.

For I3 met fields, we now read in the fields for the entire day at the start
of the day (stored in PS, SPHU, TMPU), in addition to the data for the first
timestep of the next day (stored in PS_NEXTDAY, SPHU_NEXTDAY, TMPU_NEXTDAY).
On the first timestep of the simulation, both FlexGrid_Read_I3_1 and
FlexGrid_Read_I3_2 are called to fill the *1 and *2 fields of PS, SPHU, and
TMPU. On later timesteps, FlexGrid_Read_I3_2 is called and the appropriate
time slice is selected for the *2 fields (i.e. t+3hours). If the simulation
is at hour 19 or later, the I3 fields for the first timestemp of the next
day (*_NEXTDAY) will be obtained from HEMCO.

In hcoi_gc_main_mod.F90 and in HEMCO_Config.rc the met fields read from
restart have been renamed from TMPU, SPHU, PSWET, and PSDRY to TMPU1, SPHU1,
PS1WET, and PS1DRY to reflect the fields that are read in.

NOTE: FLASH_DENS and CONV_DEPTH are still read hourly because files are
monthly, not daily. To improve I/O and read in these fields daily like the
rest of the met fields, we should split up the monthly FLASH_CTH_* files
into daily files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>